more optimization; code rearange
