+++ src/local/__init__.py
@@ -2,7 +2,7 @@ from consts import CURRENT_SYSTEM, HP
 
 
 if CURRENT_SYSTEM == HP.WINDOWS:
-    from ._winappfinder import WindowsAppFinder  # noqa
+    from .winappfinder import WindowsAppFinder  # noqa
     AppFinder = WindowsAppFinder
 elif CURRENT_SYSTEM == HP.MAC:
     AppFinder = None
